refactor: change the struct to interface and refactor the dependence.
