+++ Backend/cmd/migrate/migrate.go
@@ -4,8 +4,7 @@ package main
 
 import (
 	"github.com/Game-as-a-Service/The-Message/config"
-	"github.com/Game-as-a-Service/The-Message/domain"
-	repository "github.com/Game-as-a-Service/The-Message/service/repository/mysql"
+	"github.com/Game-as-a-Service/The-Message/service/repository"
 	_ "github.com/joho/godotenv/autoload"
 	"gorm.io/gorm"
 )
@@ -27,11 +26,11 @@ func MigrationMysql(db *gorm.DB) {
 }
 
 func Migration(db *gorm.DB) {
-	err := db.AutoMigrate(&domain.Game{})
+	err := db.AutoMigrate(&repository.Game{})
 	if err != nil {
 		return
 	}
-	err = db.AutoMigrate(&domain.Player{})
+	err = db.AutoMigrate(&repository.Player{})
 	if err != nil {
 		return
 	}
