+++ Backend/service/repository/mysql/mysqlPlayerRepository.go
@@ -2,34 +2,22 @@ package mysql
 
 import (
 	"context"
-	"time"
 
-	"github.com/Game-as-a-Service/The-Message/domain"
+	"github.com/Game-as-a-Service/The-Message/service/repository"
 	"gorm.io/gorm"
 )
-
-type Player struct {
-	gorm.Model
-	Id           int `gorm:"primaryKey;auto_increment"`
-	Name         string
-	GameId       int `gorm:"foreignKey:GameId;references:Id"`
-	IdentityCard string
-	CreatedAt    time.Time `gorm:"autoCreateTime"`
-	UpdatedAt    time.Time `gorm:"autoCreateTime"`
-	DeletedAt    gorm.DeletedAt
-}
 
 type PlayerRepository struct {
 	db *gorm.DB
 }
 
-func NewPlayerRepository(db *gorm.DB) domain.PlayerRepository {
+func NewPlayerRepository(db *gorm.DB) *PlayerRepository {
 	return &PlayerRepository{
 		db: db,
 	}
 }
 
-func (p *PlayerRepository) CreatePlayer(ctx context.Context, player *domain.Player) (*domain.Player, error) {
+func (p *PlayerRepository) CreatePlayer(ctx context.Context, player *repository.Player) (*repository.Player, error) {
 	err := p.db.Table("players").Create(player).Error
 	return player, err
 }
