+++ Backend/service/repository/playerRepository.go
@@ -1,14 +1,21 @@
-package domain
+package repository
 
 import (
 	"context"
+	"time"
+
+	"gorm.io/gorm"
 )
 
 type Player struct {
+	gorm.Model
 	Id           int `gorm:"primaryKey;auto_increment"`
 	Name         string
 	GameId       int `gorm:"foreignKey:GameId;references:Id"`
 	IdentityCard string
+	CreatedAt    time.Time `gorm:"autoCreateTime"`
+	UpdatedAt    time.Time `gorm:"autoCreateTime"`
+	DeletedAt    gorm.DeletedAt
 }
 
 type PlayerRepository interface {
