+++ Backend/tests/e2e/game_api_test.go
@@ -11,7 +11,6 @@ import (
 	"testing"
 
 	"github.com/Game-as-a-Service/The-Message/config"
-	"github.com/Game-as-a-Service/The-Message/domain"
 	handler "github.com/Game-as-a-Service/The-Message/service/delivery/http/v1"
 
 	mysqlRepo "github.com/Game-as-a-Service/The-Message/service/repository/mysql"
@@ -20,7 +19,7 @@ import (
 )
 
 var serverURL string
-var gameRepo domain.GameRepository
+var gameRepo *mysqlRepo.GameRepository
 
 func TestMain(m *testing.M) {
 	testDB := config.InitTestDB()
